Refactor effects to factory pattern: enable individual prices and display name
