+++ core/src/io/github/lonamiwebs/klooni/interfaces/IEffect.java
@@ -1,4 +1,4 @@
-package io.github.lonamiwebs.klooni.effects;
+package io.github.lonamiwebs.klooni.interfaces;
 
 import com.badlogic.gdx.graphics.g2d.Batch;
 import com.badlogic.gdx.math.Vector2;
